requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
qdrant-client>=1.7.1
sentence-transformers>=2.2.2
python-docx==0.8.11
PyPDF2==3.0.1
numpy>=1.24.0
pyjwt>=2.8.0
supabase>=2.0.0
python-dotenv>=1.0.0